# packages/contracts/openapi/cms.yaml
openapi: 3.1.0
info:
  title: Candidate Supply CMS API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /mailboxes/{id}/pause:
    post:
      operationId: pauseMailbox
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mailbox paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiGenericObjectEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
  /mailboxes/{id}/resume:
    post:
      operationId: resumeMailbox
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mailbox resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiGenericObjectEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
  /mailboxes/{id}/sync:
    post:
      operationId: requestMailboxSync
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Mailbox sync queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiGenericObjectEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/Unauthorized'
      security:
        - sessionCookie: []
          csrfHeader: []
...
components:
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    Conflict:
      description: Optimistic concurrency or duplicate relation conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
...
    Forbidden:
      description: Permission required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
    ServiceUnavailable:
      description: External provider or queue is unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiProblem'
  schemas:
...
    ApiProblem:
      type: object
      required:
        - error
        - requestId
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
        requestId:
          type: string
          minLength: 1
...
    ApiError:
      type: object
      required:
        - code
        - messageKey
      properties:
        code:
          type: string
        messageKey:
          type: string
        params:
          type: object
          additionalProperties: true
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        currentVersion:
          type: integer
          minimum: 0
          nullable: true
    GenericObject:
      type: object
      additionalProperties: true
...
    ApiGenericObjectEnvelope:
      type: object
      required:
        - data
        - requestId
      properties:
        data:
          $ref: '#/components/schemas/GenericObject'
        requestId:
          type: string
          minLength: 1
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cms_sid
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token